location "/" 일때 버그 수정과 코드리팩토링

+++ include/config/ServerConfig.hpp
@@ -15,13 +15,13 @@ public:
     private:
         friend class ServerConfig;
         map<string, LocationConfig> uriMap;
-        LocationConfig defaultLocConf;
+        LocationConfig              defaultLocConf;
     public:
         const LocationConfig &getLocConf( string uri ) const;
     };
 
     uint32_t getIp( void ) const;
-    int getPort( void ) const;
+    uint16_t getPort( void ) const;
     const vector<string> & getServerNames( void ) const;
     const LocationMap & getLocationMap( void ) const;
 private:
